Bug in strToBa fixed
The previous implementation utilized BigInteger#toByteArray which returns
a 2's complement byte array. This caused the length of the byte array to
be 17 whenever the most significant (first) byte is less than 0 so that
the sign bit is still set to 0.

This commit changes the implementation to simply loop through the String
and call Integer#parseInt